#642 Fix bug: breakable walls spawn pieces indefinitely

+++ Assets/Scripts/Mechanics/BreakableWall.cs
@@ -141,7 +141,7 @@ public class BreakableWall : SavableMonoBehaviour, IBlastable, ISetupable
         get => new SavableObject(this, "integrity", integrity);
         set
         {
-            Integrity = value.Int("integrity");
+            Integrity = value.Float("integrity");
         }
     }
 }
